requirements-test.txt
django-wkhtmltopdf
tox
flake8
pytest-django
coverage
